backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27

SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13

pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0

python-multipart>=0.0.9
email-validator>=2.1

python-jose[cryptography]>=3.3
passlib==1.7.4
bcrypt==4.0.1

fastapi-mail>=1.4

boto3>=1.34

openpyxl>=3.1

pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
